web: Show number of CoIs only on CSV tabulation
Now show the number of CoIs on the CSV version of the detailed tabulation
(where it was previously not shown).  It has been removed from the HTML
version of the tabulation in order to save space in the table.  The PI
name is still shown, but is no longer a link and is truncated rather
than replaced with an <abbr> tag when too long, since the hovertable
will make the link or <abbr> redundant.