type: Add ProposalCollection type
The new collection type has a "values_by_facility" method which iterates
over values for the given facility identifier.

This avoids the need for the personal proposals and reviews pages to
use "groupby" to organize proposals by facility.  One advantage is that
this ensures that the facilities are listed these pages in the configured
order rather than numberical facility identifier order.