type: Add ADDABLE review state
This is to be used in conjunction with reviews which a person
may add to the system.  It should not be used for any reviews
which are already present in the database.

This new state has been marked "not available" and is not returned
by the "get_options" method (now inherited from EnumAvailable)
since we do not want to offer it as a database search criterion.