web: Use auth.for_review for "Your Reviews" page
Each entry on the personal reviews page is now checked by the auth
module's "for_review" function.  This avoids the need for the view
function for this page to try to guess which reviews are editable.

Another effect of this is that authorization is checked from the point
of view of the current user, not necessarily the person whose reviews
are being listed.  So a user with administrative access will not see
links to reviews which they cannot edit, such as those for proposals of
which they are a member.