web: Extract method to prepare call defaults
This method provides the information intially displayed in the
form used to defined new calls.  Having it be a separate method
will allow facilities to provide their own defaults.  It also
takes the call type as an argument to allow it to provide
different defaults for different call types.  (This was the
reason for having the call type be part of the new call page
URL rather than a selection box on the page.)