type: Improve values_in_sorted_order method
The CollectionSortable.values_in_sorted_order method previously
assumed a total ordering and would have lost entries with the same
"key" value.  This commit updates the method so that it stores a
list of entries for each "key" value so that it can cope with cases
where the sort attributes don't specify a total ordering.  It also
adds a test case for this mix-in.