web: Add plain_group_by template filter
This is used in place of the Jinja built-in "groupby" filter on the
"your proposals" page.

The new alternative filter does not sort the data: this saves
unnecessarily attempting to sort if the data are already sorted
and allows for orderings other than those obtained by sorting by
the grouping key.  It is based on the `groupby` and `attrgetter`
functions in the Python standard library.