type: Sort ProposalCollection by start date
Now also sort proposals by the semester start date, in reverse order.
This changes the approach taken by the "values_in_sorted_order" method:
since we now need to be able to sort in both forward and reverse order,
it sorts the proposals in multiple steps as recommended by the Python
"Sorting HOW TO".  The values are now all placed in a list (which is
sorted in-place multiple times) so the method returns the list rather
than yielding individual values.