web: Check for calls with same semester / queue
Now when adding a new call, check for existing calls (of the same)
type with the same semester and queue.  This allows the problem to
be detected before attempting to add the call to the database, when
the uniqueness constraint will apply.  Also added a JavaScript
routine to indicate on the page itself whether a semester and queue
combination already exists.