web: Improve handling of datetimes from forms
Now pass datetimes to form templates as a "DateAndTime" tuple
containing the date and time as (two individual) strings.  This is
to allow unparseable input to be returned to the user for correction.
The "parse_datetime" function now accepts one of these tuples (rather than
reading from the form directly) and raises a "UserError" if parsing fails.
A corresponding "format_datetime" converts an existing datetime into a
"DateAndTime" tuple to provide the initial value for an edit form.  This
means that the template filters "format_date" and "format_time" are
no longer required, so they have been removed.