web: Configure submission notif/n by call type
The groups (if any) to which a notification is sent on the submission
of a proposal are now configured by the call type class.  The message
template no longer assumes that notifications are only sent for calls
with immediate review -- an "immediate_review" template context parameter
now specifies whether or not this applies.

Note that this constrains the order in which the enum-style classes are
defined: since BaseCallType now refers to GroupType, it must appear
after it in the module which defines them.