jcmt: Always separate ancillaries in request total
This removes the JCMTRequestCollection.get_total method's ancillary_mode
option so that it always returns "separate" entries using (instrument,
ancillary) pairs.

The JCMTRequestTotal class now extends the "namedtuple" class to add
properties which group or sum (by instrument) the instrument values.
These properties are then accessed by the HTML template.

The reason for making this change is to solve the problem of sorting the
detailed tabulation page by instrument.  Previously the template context
contained request/allocation totals made in grouped mode.  Passing these
as part of the "dynamic" component of the JSON for the row "sortinfo" data
caused the instrument dictionaries (by ancillary) to be converted to JSON.
Now the template can use the "instrument_total" property to obtain the
sorting information, restoring the proper sorting functionality.

Another advantage of the new approach is that the information for the
detailed tabulation is generated in the same way for the HTML and CSV
versions.  The template is now in full control of how the information
is presented in the HTML form.

However a disadvantage is that unavailable ancillaries are rejected
early in the "get_total" method -- the time is then allocated to the
"unknown" instrument.  This is as desired for the CSV form of the
tabulation but the HTML version was previously able to show the time
under the correct instrument even if the ancillary was unavailable.
In practise this should not be a major problem as the HTML version's
primary purpose is for use during TAC meetings, when the set of
available instruments and ancillaries should still be the same as
at the time of the call.