web: Show decision on review edit/add pages
The proposal is retrieved with the "with_decision" option for the
new review and review edit pages.  This is to allow the decision to
be shown as part of the proposal summary.  For the proposal decision
page the render_proposal_summary macro is called with
include_decision=false in order to not have the decision shown there
(since the page contains the form for editing the decision).