web: Improve display of call names
This adds a new "name_proposal" attribute to the BaseCallType.TypeInfo
to indicate whether the name relates to the call or to the proposals.
Then a new method "get_full_call_name" can produce a full name,
either singular or plural.

E.g. "regular calls for proposals" or "call for urgent proposals".