admin: Add notification after immediate review
This updates the send_call_proposal_feedback function to send an
additional notification message when sending feedback for a proposal
for an immediate review call.  For now this message is simply sent
to site administrators, but it should perhaps instead go to a group
of people configurable on a per-facility basis.

The feedback function makes use of the new get_call with_facility_code
option to retrieve information about the call rather than just fetching
the facility code (as with the old get_call_facility method).  This
allows it to determine the call type at the start of the method.
Since proposals in immediate calls are likely to be reviewed individually,
the function skips the calculation of overall ratings and the division
of the proposals into quartiles in this case.