web: Allow order of proposal sections to be set
The "proposal_view.html" template now places the main display blocks
within an "orderblocks" tag.  The ordering is then specified by a
facility view method "get_proposal_order".  The same ordering is
also used for the call edit and view templates (since they have been
arranged to match the order of the template) with the caption and
explanation word limits now shown at the end.

The original proposal ordering has been specified for the generic
facility and the current ordering specified for JCMT.

The intention of this change is that, as well as allowing facilities
to customize the order in which proposal sections are shown, it will
also be useful in allowing specific parts of a proposal to be extracted
when generating a PDF view of it.