web: Divide science and technical case into blocks
Now the science_case and technical_case blocks are subdivided
into three more blocks: intro, content and extra.  The "intro"
block contains the title, the "content" the actual text or
PDF pages and the "extra" block can be overridden by templates
to add additional features.

These blocks are enclosed in an "orderblocks" structure in
order to give control over the blocks to be shown.