web: Allow a proposal to be viewed in sections
This adds a new section-based route for viewing a proposal where
each section can either be HTML (rendered via a template) or a PDF
file to include (e.g. an uploaded science case).

The sections are returned via a kind of multi-part response.  This is
currently implemented using pickle for simplicity.  (A true MIME multi-part
response would of course be preferable.)