publication: Only decode XML text with Python 2
In Python 3, it appears that xml.etree.ElementTree returns text
as unicode, so we should not attempt to decode it.  In Python 2 it
returns ASCII (byte strings) where possible or unicode (otherwise)
so we need to check the type before attempting to decode.  Also now
decode as ASCII rather than UTF-8 in this case.