jcmt: Partly revert affiliation assignment method
This partially reverts 0e7c34717e56af7cf36125ba300cdeb5773227d8
to restore shared affiliations for proposal CoIs.  However this
is now only applied when a CoI inherits the PI's affiliation
(i.e. the CoI affiliation is SHARED or EXCLUDED and the PI
affiliation is SHARED) and not when the CoI affiliation is
SHARED but the PI has a non-SHARED affiliation.

Requested by JTD.