web: Close shown tips when opening a new one
Now when the "formtip" question mark is clicked, if the corresponding
tip is not currently visible, all tip boxes are closed and then the
selected box is shown.  This ensures that only one box will be open
at a time, in order to prevent the screen becoming cluttered with
them.  To do this we now have to test if the box is currently visible
rather than simply calling "toggle" on it.