email: Separate out message preparation function
The routine which prepares an email message for sending has been
separated out into a separate function so that it can be run
by the test suite (without actually sending email).

Note: the message date time zone is currently processed incorrectly
and the unit test has been set up to expect this for now.  (And depends
on the time zone being HST.)