email: Correct timezone handling for date header
For Python 2, since we have datetime objects in UTC, we should use
'calendar.timegm' in place of 'time.mktime'.  But there is a problem
with 'email.utils.formatdate' in that it writes the timezone for UTC
as "-0000" instead of "+0000", so we need to detect and correct that.

For Python 3, we can use 'email.utils.format_datetime' instead,
which works correctly but requires the datetime object to include
a timezone.  Therefore we must insert the UTC timezone info, but
we can only do that in Python-3-specific code since Python 2 doesn't
include 'datetime.timezone' and we don't otherwise have a direct
dependency on pytz.