jcmt: Move calculation logic to SCUBA-2 ITC module
This is to simplify the Hedwig ITC interface class and make the
SCUBA-2 ITC module easier to use on its own.

Also no longer rely on SCUBA2ITCError having a "message" attribute.
We now expect the ITC to convert division by zero and negative square
root exceptions to instances of this class.