jcmt: Add map sampling option to SCUBA-2 ITC
This replaces the check-box for the matched filter.  The default
sampling option is now a map-type-specific recommendation.  Other
options are matched filter and custom pixel sizes.  The calculator
interface version has been updated to 2 to account for this change.
Currently the recommended pixel sizes are just 6.5" and 4" for
all plain SCUBA-2 observing modes, and 12" (both wavelengths)
for observing with POL-2.

Requested by JTD at the meeting about this ITC on 3/2/17.