scripts: Update interface for poll task selection
Now rather than specifying "all" or one particular poll task, it is
possible to select multiple tasks with options (e.g. "--figure --pdf").
If no tasks are specified in this way, then they will all be run
except for those disabled with "no" options (e.g. "--no-email").

A "--dry-run" option has also been added.  At the moment none of the
poll functions support a dry run option, so when this is specified
they are not run at all.  However the intention is that some or all
functions will gain a useful dry run mode in the future.