db: Add method to merge person profiles
This is based on the previous code which was part of the routine
for accepting an invitation when logged in as a fully registered
user.  The "use_invitation" method now calls the new routine.

This currently requires that we know all the tables which could refer
to person records so that we can update / delete them as appropriate.
It ought to be possible to extract that information from the SQLAlchemy
metadata instead.