web: Query reviewer invitation for reminder
Since commit 32c0810b7a, @with_review has been used to obtain
review information when sending a reminder message.  However this
did not include the details of the previous invitation token.
This commit adds a "with_invitation" argument to the "with_review"
decorator so that this information can be restored and reminder
messages can repeat the existing token if it is still valid.