db: Rename invitation and token methods
Rename invitation and token creation methods to begin with
the prefix "issue" (as opposed to a mixture of "get" and "add").
These match "use" methods which make use of the corresponding
tokens.  This is because these are methods designed to write
to the database, so they should start with a verb which indicates
that.