web: Number facility's target tools sequentially
Now assign sequential identifiers to the target tools of each
facility.  (We do not have a database table of target tools,
as there is for calculators, as this has not yet been necessary.)
Previously they were all assigned identifier 0, so multiple
target tools would overwrite each other in the facility's tools
dictionary.