db: Add dedicated proposal co-member search method
This is now used by the "auth" module instead of the previous
co-membership constraints to the "search_member" method. The
reason for this is that the results from the new method are
more amenable to caching: we can get a list of people who
are proposal co-members with a given person and include additional
information about the given person's role in that co-membership
(namely whether or not they are an editor of the proposal).
This was not possible with the "search_member" method because
it only returned information about one member, allowing the
other to be specified as a constraint only.