astro: Let parse_source_list return an object list
The parse_source_list function now has an option
"as_object_list" which can be used to request a list of
TargetObject instances rather than a TargetCollection.
This is so that target tools in upload mode do not
have to convert from decimal degrees back to objects,
which was inefficient because these objects were used
as an intermediate step in parsing the source list.