web: Avoid get_flashed_messages clearing session
Flask's "get_flashed_messages" stores the messages it retrieves in the
request context.  This means that if it is called again (as is done by
the template when displaying them) the cached messages are returned and
it does not clear them from the session again.  This could result in
messages, such as the logged out message, never going away.

This commit therefore retains the flashed messages when clearing the
session instead by preserving the "_flashes" session entry.  This assumes
we know the internal name used by Flask for the flashed messages, but
if this fails, it will not cause a disaster -- we may just loose some
messages such as the logged out message.  The alternative would be to
specifically clear only our own entries from the session, but this would
risk not clearing any subsequently added entries, which could potentially
be important to clear.