web: Use list comp. instead of gen. for tab. rows
For the CSV download of the tabulation, _get_proposal_tabulation_rows
loops over proposals and yielded a generator for the affiliations of
each.  We zip the result of this method, which in Python 2 causes every
proposal's entry to be yielded first, before any are output.  This seems
to create a closure with the proposal loop variable so that, at the
time the rows are printed, each gets a copy of the affiliations for the
last proposal in the list.  This could be fixed by using izip instead
of zip in Python 2, but for now the generator has simply been changed
to a list comprehension.  This is fully evaluated in advance so does
not have the problem described above.