web: Allow reviews to be saved in a draft form
This adds a checkbox to the end of the review form to allow the reviewer
to indicate when they consider the review complete.  When the box is not
checked, the review is stored with a state of "PREPARATION".  The review
should then be considered as not done (because the state is not "DONE")
elsewhere in the system.