jcmt: Add redshift to the heterodyne ITC
Moved the frequency input to the "source and conditions" section,
and renamed it "rest frequency".

The JavaScript ITC interface performs a redshift calculation to
determine the tuning range of the receiver in terms of the rest
frequency.  However the calculator itself uses the heterodyne ITC
module to perform the redshift calculation.