jcmt: Add transition lines to the heterodyne ITC
This makes the line catalog from the ITC module available to the
JavaScript code managing the ITC page as JSON via a new custom route.
This is so that the browser can cache the JSON representation of the
catalog, avoiding the need for it to be sent with each request to
the ITC.

The frequency can now either be specified as a transition line
(species name and transition name) or as a rest frequency.  When a
species is selected, the ITC should populate the list of transitions
and enable only those available at the current redshift.  Selecting
a transition will have it shown in the (disabled) frequency box.
If "Other" is selected then the frequency box will be enabled.