web: Simplify code for reviewer grid page
Now use lists of reviewer roles rather than having "primary"
and "secondary" variables.  This simplifies the code (and would
allow more than two roles to be shown in the grid if that were
ever required).  Now the role information is passed to the
template, so it can indicate what the two roles being assigned
are.  Also pass the template a {person: review_present} dictionary
rather than two separate lists.