web: Avoid conflicting assignments in reviewer grid
This adds two more checks when the reviewer assignment grid is
submitted: that the same person has not been assigned to multiple
roles via the grid and that no-one who already has a "conflicting"
review is given an assignment.  In the case of the committee
primary and secondary reviewer roles, the committee "other" role
is considered to be conflicting.

The web interface has been updated to show cross marks where a
review cannot be assigned due to a conflict (and dashes where a
"unique" role already has a submitted review).  A JavaScript routine
to disable conflicting form controls has been added.  The documentation
for this page has been updated to describe the non-control elements
which may appear in the grid.