web: Simplify message handling in target tools
Now allow for a "message" (indicating an error dealing with the
user's input) in all tool modes (i.e. add it to "proposal" mode).
In "proposal" mode, if a message is set, then it is raised via
ErrorPage unless the tool indicates that it can handle the message
in its template via the class variable "proposal_message_error".

Also now apply the don't-overwrite-with-None check for messages
in the BaseTargetTool class so that individual target tools do not
need to worry about this.