db: Add primary_email option to add_person method
This is to allow the person record and their primary email address
record to be created in the same transaction, so that, if the
email address fails to validate, the person record is not committed.
The web interface has been updated to use this new option rather than
calling "add_person" followed by "add_email".