web: Allow calculations for technical assessments
The BaseCalculator class has been updated to allow for saving of
calculations to reviews (in addition to proposals).  The BaseReviewerRole
class now has an info field "calc" to specify which roles can have
calculations -- this has been set to true just for technical assessments.

The calculation management page can now be invoked for calculations attached
to a review -- a new route has been added for this.  Calculations are now
displayed on the review edit and proposal reviews pages.