web: Update logic for rows on proposal reviews page
Previously the main review text row was only shown if the text
itself was not none.  However this would mean that any information
added by a child template, or any calculations, would be hidden
if the text were none.  Now the row content is pre-rendered (using a
block assignment) so that we can check whether it really is empty.

In addition, calculations have been moved to their own row
of the table to improve the layout.

Inheritance blocks have been removed from the calculation and
note rows. (If we wish to override these in a child template,
then they also will need to be pre-rendered for the same reason
as the review text.)

(As a result, templates need to check if the review text is none
or not before attempting to format it, because this is done before
we decide whether to show the row or not.)