web: Add dedicated routes for saved calculations
Now use dedicated routes rather than query parameters to view calculations
which have been saved for a proposal or review.  This is so that we can use
the "require_auth" decorator on these new routes in order to have the user
be prompted to log in if necessary.  (Previously we would check their
authorization but could not invoke the log-in mechanism.)  In order to
implement this, the view functions are saved in the CalculatorInfo tuples
belonging to each facility.  The new saved calculation view functions
retrieve the relevant calculation and then invoke one of these saved
calculator view functions.

The new route scheme is simpler, as we can generate the URLs from just
the proposal or review and calculation identifiers.  It also allows us to
make use of the "with_proposal" and "with_review" decorators.