web: Allow proposal members to be reordered
This adds a new JavaScript library file "table_row_drag.js" which
allows the rows of a table to be dragged to change their order.
When a row is dropped, the "sort_order" values of the rows are
updated so that they match the order in which the rows are now shown.