db: Combine proposal text retrieval methods
Adding a "with_text" argument to the "search_proposal_text" method
has allowed the separate implementations of the "get_proposal_text"
and "get_all_proposal_text" methods to be removed.

This also allowed the old "ProposalText" type to be replaced with
a combined type based on "ProposalTextInfo", which has itself been
removed.