db: Use link tables for PDFs, figures and text
This adds a new set of tables which are used to link PDFs, text and figures
to proposals (and figures to reviews).  The proposal (or review) identifiers
have been removed from the tables where the large data objects are stored.
They are now only present in the corresponding link tables.

The purpose of this change is to allow these large database rows to be
shared between multiple proposals, for example when a proposal is copied.
This will allow copying to be a less expensive operation than it would
be if it involved immediately copying all of the PDF uploads, figures
and online-edited text.

When the underlying large data object is changed, a new record is inserted
into the database and the associated link updated to refer to it.  This
should prevent the version which is linked to any other proposals from
changing.  Update operations performed in this way are then followed
by a step to delete any records with no references left.