web: Add support for review acceptance
Now for reviewer roles requiring acceptance, make authorization
dependent on having accepted the review.  If the review has not
been accepted or rejected yet, redirect to an acceptance page.
This page shows the proposal information, abstract, categories
and member list and asks the reviewer to indicate whether or not
they would have a conflict of interest in providing a review.
The URL for the page triggering the redirect is stored as the
referrer so that, if the review is accepted, the user can
be redirect back there.  Otherwise they are redirect to the
review's information page.

When searching for proposals with a given person as reviewer,
rejected reviews are now excluded.