db: Add reviewer acceptance table
Note that this is not linked to the reviewer table directly, but instead
is identified in the same way as reviewers: by person, proposal and role.
This is so that acceptance records can be retained for informational
purposes if the corresponding reviewer records are deleted.

The acceptance records also duplicate the accepted flag from the reviewer
table and include the date they were last updated.