jcmt: Add role settings for peer reviews
This adds "peer_rating" and "peer_expertise" as columns of the
"jcmt_review" table.  They take values from two new enum-style
classes.

New classes were used to keep these values separate from the
existing (TAC) reviewer expertise, in order to allow them to
have different values and be maintained individually in the
future.  The rating and weight values from the new classes
are used in the calculation of the proposal's overall rating.