web: Adjust redirect after accepting invitation
Previously a proposal took precedence over reviews when determining
where to redirect a user after they have accepted an invitation.
However this is not helpful in the case of an unregistered member
of a proposal being assigned peer reviews.  So if the invitee has
a mixture of proposals and reviews, they are now redirected to their
profile page, which has links to their proposals and review pages.
If they have multiple proposals, or multiple reviews, then they are
now redirected to their proposal or review list respectively.

In addition the "first_view" URL parameter was removed from the
redirect to a particular proposal, as that now causes the copy
report to be show, if present.

The logic for determining the redirect target has been extracted
into a new private method so that it can be tested in the test suite.