web: Add log_in_for arg. to password reset pages
This is to allow us to track which page the user wished to log in
for when they were prompted to reset their password.  If no "log_in_for"
argument is present at the end of the process, the home page is used,
in order to avoid the log in page trying to redirect the user back to
the password reset page once they have logged in, since that causes
an "already logged in" error.