jcmt: Reverse dual polarization option in Het. ITC
This changes the "dual polarization" option to the logical reverse
of this, "separate polarizations".  The default is "false" (for HARP,
as it only observes one polarization).  This means that for receivers
which do observe two polarizations, the default is now to assume they
will be reduced together, as this is the more normal usage.