web: Set calc. default review based on proposal
If a calculation for a particular proposal is viewed and the user
can edit a review of the proposal, that review should now be the
default review to which the calculation can be saved.

Also when viewing a proposal calculation with an unknown code
(because the user can not edit the proposal), check for reviews
of the proposal to try to find the code.  (There are still
many circumstances when the code will not be available -- maybe
we should just query the database directly to find it.)

Additional documentation was added to the calculator view method
to explain the different sets of proposal and review identifiers
which it tracks.