web: Fetch proposal and review for calc. nav links
Previously calculators could only show the proposal code and reviewer
role of the current proposal / review if a proposal or review was
editable by the user.  Now the proposal calculation and review
calculation view functions pass in the parent proposal and review.
If these are not given (e.g. when the calculation is adjusted and
re-run) then the proposal and reviewer are fetched from the
database when required.