type: Add "viewer" review group type
Membership of this group grants permission to view proposals and
feedback.  Also, for the time being, it grants permission to view
reviews, so that members of the group can navigate to proposals via
review process pages.  (If an alternative navigation system is added,
this permission could be reconsidered.)

Feedback view permission has been added as an extra piece of
information in the group information table and the "group_view"
template has been updated to show a more complete set of permissions.