web: Add initial version of review allocation page
This page will allow the RA distribution of targets associated
with a call's proposals to be plotted.  It has two routes: one which
shows the page and a second which provides dynamic information
in JSON format.  The initial page load includes fixed information
as well as the current version of the dyamic information in
data attributes.

The page uses Chart.js to create the plot.  The current version
draws a stacked bar chart by proposal, but the intention is for the
JavaScript layer to allow the data to be categorized in a more
useful manner.