Use request system to create copies of proposals
The copying process can take some time when applied to complete
proposals, which seems to have led to people making multiple
copies since the system did not indicate that progress was occurring.
Now the request status page should load immediately and redirect
the user to the new proposal once copying is complete.

Copying a proposal (as currently implemented) requires a Flask
request context to identify the copier and build the URLs included
in invitation emails.  Therefore a method to obtain a Flask
application (with caching) has been added to the hedwigctl script.
The base URL must now be included in the configuration file
to allow URLs to be constructed in this context.