config: Construct facilities in get_facilities
The get_facilities function now constructs facility view objects
and their associated calculators and target tools rather than
simply returning a list of facility classes.  The reason for this
change is that this functionality does not seem really to belong
in the web application and facility blueprint construction
functions.

If no database usage (to get identifiers) is desired, the empty tuple
can be given for the database control object, with the default value
of None indicating that get_database should be used automatically.
Note that this is the reverse of the normal Hedwig convention where
None could be passed to override an empty tuple default.  However in
this case, using no database should be a very rare case (e.g. from the
test suite) and this function could conceivably be called from another
function where db defaults to None -- it would be unhelpful for the
no-database case to be triggered accidentally.  Additionally dummy
identifiers are numbered from 9000000 to allow them easily to be
distinguished from regular identifiers, or the random identifiers
used by the test suite.