web: Pass facilities to invitation_token_accept
This allows the view function to find facilities (by ID) in this
dictionary, rather than having to look up the facility code
in the database.  The previous approach could have caused problems
when an invited user had a proposal or review for a facility
which is no longer in use -- that facility would not have
a blueprint loaded so URLs for it could not be constructed.