config: Improve caching in configuration functions
This adds a new decorator class MemoCache which can be used to
cache the results of some of the configuration functions, rather
than having them use global variables.  This class interacts
with database control objects by using their new "_mem_id"
attribute as a cache key.

The advantage of the new system should be that, for functions
which take parameters, such as a facility specification or database
control object, then an appropriate cached result can be returned,
rather than the most recent result if no parameters are given.