web: Add auth function for call review by proposal
Like the for_person_member and for_person_reviewer methods,
this constructs a dummy record based on available information
to be used for authorization.  In this case, only the queue_id
value is required, so a dummy call record can be constructed
from the information included with a proposal and passed to
the for_call_review function.  This avoids needing to fetch
the call record from the database specifically for authorization.