web: Pass current_user to route wrapper functions
The auth decorators now pass the current_user object as the first
argument to route wrapper functions.  Another decorator
"with_current_user" performs this task for routes which do
not require authentication.

This arrangement has the auth decorators encapsulating the
logic required to obtain the current_user object.  (In this
case, extracting it from the flask_g object.)  From this point
it can just be passed on normally.