db: Allow aff/n settings to be overridden by call
This adds hidden and type columns to the affiliation_weight table
so that these affiliation settings can be overridden for a given
call.  Now when searching affiliations including weights for a
call, the hidden and type values are coalesced from the new
columns and the affiliation table, unless with_weight_separate
is specified, in which case the values from the affiliation_weight
table are returned separately.