email: Override header encoding to QP in Python 2
In Python 2 it seems that the header "maxlinelen" parameter may
not be fully applied if email.charset.Charset.header_encode
decides to use base64, since it does not pass a "maxlinelen"
parameter to email.base64mime.header_encode which has a default
of 76.  This can be avoided by overriding the character set
settings for "utf-8" to use only quoted-printable encoding
for headers.