email: Make is_valid_email check more specific
The pattern used by this function to check email addresses has been
updated to try to look for the basic form described in RFC2822.
A new test script has been added for this function.

Additionally an update script is provided to check whether
existing database entries match the new pattern.