email: Update email message preparation function
Now use the "EmailMessage" API on Python 3.  For Python 2, email
messages are prepared as before, but we construct address headers
(to and from) manually to ensure the email addresses in these
headers are not encoded.  For simplicity the names in these headers
are quoted whether needed or not, unless they are non-ASCII, in
which case they are encoded as quoted printable.