Rename invitation acceptance to application
In the user interface, instead of referring to accepting an invitation,
call this applying the invitation to an account.  This is to avoid
confusion.  (Using an invitation code is a technical operation
in the profiles and permissions system giving access to something,
and should not be misunderstood as agreement to take part in
something, such as a proposal or review.)