web: Single update box & state selection for req.
The user request status page will now show a single update check box
for "resettable" entries, rather than separate boxes for resetting
processing and expiry.  The page now has a state selection control,
like other status pages, and this now includes the option to discard
entries.

This is intended to simplify the page, whilst also making the state
selection more flexible.  However the view function now needs
to check that the selected state transitions are appropriate.  It
does this using a new method RequestState.is_resettable via its
"state_new" argument.  The initial implementation of this method
simply lists expiry-related states which can be reset to READY
and other processing states which can be reset to NEW or DISCARD.
However it could be extended in future, e.g. to allow requests
with expiry errors to be discarded.