web: Allow multiple links for clash tool results
This renames the facility make_archive_search_url method to
make_archive_search_urls.  It should now return a list of
"Link" namedtuples so that the link text can be provided
for each URL.  The CSS was adjusted to add padding
between <ul> and <nav><ol> when nested inside another list,
as in the clash tool results page.