web: Add validation of previous proposals for CR
For a continuation request, there will need to be a single
entry marked as "continuation" on the previous proposals
list, in order to indicate which project is to be extended.
Therefore add a more specific check to the proposal validator
for continuation requests and update the notes on the
previous proposal list editing page.