jcmt: Allow options to be altered with time alloc.
This adds a new database table "jcmt_alloc_options" with the same
columns as "jcmt_options" which is used to store options selected
at the time of entering a time allocation on the decision page.
This is to allow the options to be corrected as needed before approved
proposals are exported from the system.  The allocation options are
also shown on the proposal feedback page.