db: Add extra arguments to search_prev_proposal
This is to allow searching by the continuation flag.  Also a new
'with_publications' argument controls whether or not the list of
associated publications should be included.  This defaults to true
to retain the previous behavior.  A new unit test has been
added to check the operation of this method.